moving traits around to a different place
